+++ Frontend/src/pages/DiaryPage.css
@@ -1,5 +1,5 @@
 .diary-page {
-  background: linear-gradient(135deg, #ffeaa7 0%, #fab1a0 100%);
+  background: white;
   min-height: 500px;
   padding: 20px;
   border-radius: 15px;
