+++ Frontend/src/pages/Header.css
@@ -11,6 +11,18 @@
   border-radius: 15px;
 }
 
+.mood-selector {
+  background: rgba(255, 255, 255, 0.2);
+  color: white;
+  border: 1px solid rgba(255, 255, 255, 0.3);
+  padding: 5px 10px;
+  border-radius: 15px;
+  font-size: 11px;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  backdrop-filter: blur(5px);
+}
+
 .header::before {
   content: '';
   position: absolute;
@@ -46,6 +58,7 @@
 
 .logo {
   display: flex;
+  flex-direction: column;
   align-items: center;
   gap: 15px;
 }
@@ -64,14 +77,17 @@
   50% { transform: scale(1.05); }
 }
 
+.logo-text {
+  margin: 30px;
+}
+
 .logo-text h1 {
   font-size: 24px;
   font-weight: bold;
-  margin: 0;
   text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
 }
 
-.date {
+.now-status {
   font-size: 14px;
   opacity: 0.9;
   margin: 5px 0 0 0;
