+++ Frontend/src/pages/MiniRoom.css
@@ -1,0 +1,8 @@
+/* MiniRoom.css */
+.miniroom {
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  background: white;
+}
