+++ Frontend/src/pages/Navigation.css
@@ -1,14 +1,17 @@
 /* Navigation.css */
 .navigation {
-  background: white;
+  /* z-index: -1; */
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  background: rgba(255, 255, 255, 0);
   border-radius: 15px;
-  box-shadow: 0 5px 20px rgba(0, 0, 0, 0.1);
-  border: 2px solid #e0e0e0;
+  /* box-shadow: 0 5px 20px rgba(0, 0, 0, 0.1); */
+  /* border: 2px solid #e0e0e0; */
   overflow: hidden;
 }
 
 .nav-container {
-  display: flex;
   justify-content: space-between;
   align-items: center;
   padding: 0;
@@ -16,23 +19,25 @@
 
 .nav-tabs {
   display: flex;
-  flex: 1;
+  flex-direction: column;
+  gap: 8px;
+  background-color: rgba(0, 0, 0, 0);
 }
 
 .nav-tab {
+  border-radius: 0 18px 18px 0;
   position: relative;
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 5px;
-  padding: 15px 12px;
-  background: none;
+  padding: 8px 12px;
+  background: white;
   border: none;
   cursor: pointer;
   transition: all 0.3s ease;
   flex: 1;
   min-width: 0;
-  border-radius: 0;
   overflow: hidden;
 }
 
@@ -67,17 +72,11 @@
 .nav-icon {
   font-size: 20px;
   display: block;
-  animation: bounce 0.5s ease-in-out;
+  
 }
 
 .nav-tab.active .nav-icon {
   animation: pulse 1s infinite;
-}
-
-@keyframes bounce {
-  0%, 20%, 50%, 80%, 100% { transform: translateY(0); }
-  40% { transform: translateY(-3px); }
-  60% { transform: translateY(-2px); }
 }
 
 @keyframes pulse {
@@ -118,7 +117,6 @@
   display: flex;
   gap: 5px;
   padding: 0 15px;
-  border-left: 1px solid #e0e0e0;
 }
 
 .action-btn {
@@ -148,6 +146,16 @@
 .logout-btn:hover {
   background: linear-gradient(135deg, #ff7675, #d63031);
   color: white;
+}
+
+.avatar-emoji {
+  font-size: 60px;
+  background: rgba(255, 255, 255, 0.2);
+  padding: 15px;
+  border-radius: 50%;
+  border: 3px solid rgba(255, 255, 255, 0.3);
+  backdrop-filter: blur(10px);
+  animation: bounce 2s infinite;
 }
 
 /* 반응형 */
